+++ light.h
@@ -9,7 +9,7 @@
 
 class Light {
   public:
-    float red, green, blue;
+    long double red, green, blue;
     virtual ThreeDVector* get_light_direction_from(ThreeDVector* postion) = 0;
     virtual Ray* get_shadow_ray(ThreeDVector* postion) = 0;
     virtual vector<Ray*> get_shadow_rays(ThreeDVector* position, int sample_size) = 0;
