+++ sphere.h
@@ -8,8 +8,8 @@ using namespace std;
 class Sphere: public Surface {
   public:
   	ThreeDVector* center;
-  	float radius;
-  	Sphere(ThreeDVector* center, float radius);
+  	long double radius;
+  	Sphere(ThreeDVector* center, long double radius);
   	bool hit(Ray* _ray, Record* record);
   	ThreeDVector* get_normal(ThreeDVector* surface_point, ThreeDVector* view_vector);
   	void apply_transformation(Eigen::Matrix4f transformation);
